Package example config in container
This allows the example usage in the README to actually work, an allows
the image to be used on its own instead of just as a base image.

We remove the STS header just because it seemed presumptuous as a
default. You should absolutely add this header if you can.

Also we should fix the image URL in the usage example, as that was
_also_ wrong...